fix: Track `audio_format` is now a property and not a field
Use some regex query on path if you wish to query for it. For now, I'd like to keep track properties like bitrate, audio_format, etc. as generated properties since they require a data migration. I'd like to do that data migration all at once if possible.
